Separate the common code for a person's proxy images
